+++ CHANGELOG.md
@@ -16,6 +16,18 @@ commits since the beginning of this repository.
 ### Added
 ### Removed
 
+## [5.7.172]
+
+### Added
+
+- The ability to use tracing with Sentry. Big thanks to @karuta0825. There is an example in the `examples` directory.
+
+### Changed
+
+- Update Sentry Java SDK to 5.7.1.
+- Various library updates.
+- Update github workers cache to v3
+
 ## [5.7.171]
 
 ### Changed
@@ -256,7 +268,8 @@ commits since the beginning of this repository.
   compatible with Sentry 10.0.1 and below. If you wish to use those
   versions, please continue to use sentry-clj 1.7.30.
 
-[Unreleased]: https://github.com/getsentry/sentry-clj/compare/5.7.171...HEAD
+[Unreleased]: https://github.com/getsentry/sentry-clj/compare/5.7.172...HEAD
+[5.7.172]: https://github.com/getsentry/sentry-clj/compare/5.7.171...5.7.172
 [5.7.171]: https://github.com/getsentry/sentry-clj/compare/5.6.170...5.7.171
 [5.6.170]: https://github.com/getsentry/sentry-clj/compare/5.6.169...5.6.170
 [5.6.169]: https://github.com/getsentry/sentry-clj/compare/5.6.166...5.6.169